fix: minor fixes + refactor
