+++ .vscode/settings.json
@@ -43,4 +43,6 @@
       "mode": "auto"
     }
   ],
+  "editor.formatOnSave": true,
+  "editor.defaultFormatter": "dbaeumer.vscode-eslint",
 }
