+++ apps/directorate-of-equality-web/src/components/list-page/TabsContent/TabContent.css.ts
