+++ apps/directorate-of-equality-web/src/components/list-page/TabsContent/TabContent.tsx
@@ -5,7 +5,7 @@ import { Text } from '@dmr.is/ui/components/island-is/Text'
 import { Table } from '@dmr.is/ui/components/Tables/Table/Table'
 
 import { Paging } from '../../../gen/fetch'
-import { type Case, DETAIL_FIELDS } from '../constants'
+import { type Case, DETAIL_FIELDS } from '../../../lib/constants'
 import * as styles from './TabContent.css'
 
 import { useQuery } from '@tanstack/react-query'
@@ -59,7 +59,7 @@ export const TabContent = ({
         columns={columns}
         data={data}
         loading={isLoading}
-        getRowHref={(row) => `/mal/${row.id}?type=${row.type}`}
+        getRowHref={(row) => `/mal/${row.id}`}
         getRowExpanded={
           expandable ? (row) => <ExpandedRow row={row} /> : undefined
         }
