+++ apps/directorate-of-equality-web/src/components/list-page/TabsContent/TabsContainer.tsx
@@ -18,10 +18,10 @@ import {
   ReportStatusEnum,
 } from '../../../gen/fetch/types.gen'
 import { useReports } from '../../../hooks/useReports'
-import { type Case, COLUMN_STATUS, COLUMNS } from '../constants'
+import { type Case, COLUMN_STATUS, COLUMNS } from '../../../lib/constants'
 import { CreateEqualityReportDrawer } from '../CreateEqualityReportDrawer'
 import { CreateSalaryReportDrawer } from '../CreateSalaryReportDrawer'
-import { ReportFilter } from '../Filter/ReportFilter'
+import { ReportFilter } from '../FilterContent/ReportFilter'
 import { TabContent } from './TabContent'
 
 import { type ColumnDef } from '@tanstack/react-table'
