+++ apps/directorate-of-equality-web/src/containers/report/ReportSidebarContainer.tsx
@@ -1,20 +1,26 @@
+'use client'
+
+import { useSuspenseQuery } from '@tanstack/react-query'
+
 import { EmployeeSelect } from '../../components/report/report-sidebar/EmployeeSelect'
 import { ReportFormStepper } from '../../components/report/report-sidebar/ReportFormStepper'
 import { ReportSidebar } from '../../components/report/report-sidebar/ReportSidebar'
 import { ReportStatusSelect } from '../../components/report/report-sidebar/ReportStatusSelect'
-import { ReportStatusEnum } from '../../gen/fetch'
+import { useTRPC } from '../../lib/trpc/client/trpc'
 
 type ReportContainerProps = {
   id: string
-  status: ReportStatusEnum
 }
 
-export function ReportSidebarContainer({ id, status }: ReportContainerProps) {
+export function ReportSidebarContainer({ id }: ReportContainerProps) {
+  const trpc = useTRPC()
+  const { data: report } = useSuspenseQuery(trpc.reports.getById.queryOptions({ id }))
+
   return (
     <ReportSidebar>
       <EmployeeSelect />
-      <ReportStatusSelect reportId={id} status={status} />
-      <ReportFormStepper status={status} />
+      <ReportStatusSelect reportId={id} status={report.status} />
+      <ReportFormStepper status={report.status} />
     </ReportSidebar>
   )
 }
